+++ semgrep-core/parsing/Parse_csharp_tree_sitter.mli
@@ -1,2 +1,2 @@
 
-val parse: Common.filename -> AST_generic.program
+val parse: Common.filename -> AST_generic.program Parse_tree_sitter_helpers.result