+++ semgrep-core/parsing/Parse_go_tree_sitter.mli
@@ -1,2 +1,2 @@
 
-val parse: Common.filename -> Ast_go.program
+val parse: Common.filename -> Ast_go.program Parse_tree_sitter_helpers.result