+++ semgrep-core/parsing/Parse_ruby_tree_sitter.mli
@@ -1,2 +1,2 @@
 
-val parse: Common.filename -> Ast_ruby.program
+val parse: Common.filename -> Ast_ruby.program Parse_tree_sitter_helpers.result