+++ semgrep/tests/public_repos.py
@@ -1,7 +1,7 @@
 import pytest
 
 
-def xfail_repo(repo, *, reason=None):
+def xfail_repo(repo, *, reason=""):
     return pytest.param(repo, marks=pytest.mark.xfail(reason=reason, strict=True))
 
 
@@ -277,11 +277,11 @@ FAILING_REPOS = [
             "repo": "https://github.com/highcharts/highcharts",
             "languages": ALL_LANGUAGES,
         },
-        reason="javascript parse error, probably due to ?. regression",
+        reason="javascript parse error, probably due to '?.01' regression in minified file",
     ),
     xfail_repo(
         {"repo": "https://github.com/OWASP/NodeGoat", "languages": ALL_LANGUAGES},
-        reason="javascript parse error, probably due to ?. regression",
+        reason="javascript parse error, probably due to 's.behaveLikeLine?.8:1' regression in minified file",
     ),
     xfail_repo(
         {"repo": "https://github.com/coinbase/gtt-ui", "languages": ALL_LANGUAGES},
@@ -344,11 +344,19 @@ FAILING_REPOS = [
     ),
     xfail_repo({"repo": "https://github.com/Homebrew/brew", "languages": ["ruby"]}),
     xfail_repo({"repo": "https://github.com/gitlabhq/gitlabhq", "languages": ["ruby"]}),
-    xfail_repo({"repo": "https://github.com/coinbase/coinbase-pro-node", "languages": ["javascript", "typescript"]},
-                reason=("Failure to parse typescript"),
+    xfail_repo(
+        {
+            "repo": "https://github.com/coinbase/coinbase-pro-node",
+            "languages": ["javascript", "typescript"],
+        },
+        reason=("Failure to parse typescript"),
     ),
-    xfail_repo({"repo": "https://github.com/bkimminich/juice-shop", "languages": ["javascript", "typescript"]},
-                reason=("Failure to parse typescript"),
+    xfail_repo(
+        {
+            "repo": "https://github.com/bkimminich/juice-shop",
+            "languages": ["javascript", "typescript"],
+        },
+        reason=("Failure to parse typescript"),
     ),
     xfail_repo(
         {"repo": "https://github.com/dropbox/changes", "languages": ALL_LANGUAGES},